% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_last_fire.R
\name{make_last_fire}
\alias{make_last_fire}
\title{Make last fire layer from fire history polygons}
\usage{
make_last_fire(
  fires_lyr,
  date_field = "ignition_date",
  parallel = FALSE,
  prop_cores = 0.5,
  max_cores = 124
)
}
\arguments{
\item{fires_lyr}{sf object containing fire polygons.}

\item{date_field}{Character. Name of field indicating the fire start dates.}

\item{parallel}{Logical. Run in parallel?}

\item{prop_cores}{Numeric. Proportion of available cores to use if \code{parallel} == TRUE.}

\item{max_cores}{Integer. Maximum number of available cores to use if \code{parallel} == TRUE.}
}
\value{
sf object with polygons representing the last fire in an area.
}
\description{
Make last fire layer from fire history polygons
}
